Fully Implementing encryption

+++ db/migrate/20190711012827_create_wallets.rb
@@ -2,7 +2,7 @@ class CreateWallets < ActiveRecord::Migration[5.2]
   def change
     create_table :wallets do |t|
       t.references  :user
-      t.float       :balance, null: false, default: 0
+      t.integer     :balance, null: false, default: 0
       t.string      :currency_type, null: false, default: 'usd'
       t.timestamps
     end
